배포 전 사전작업
