+++ src/app/(main)/shops/[shopId]/page.tsx
