+++ src/app/(main)/shops/[shopId]/register/components/ImageUploader.tsx
