+++ src/components/GNB/GlobalHeader.tsx
@@ -10,131 +10,133 @@ import clsx from "clsx";
 import { useUserAlerts } from "@/hooks/api/useAlertService";
 import { GetListQuery } from "@/types/common";
 import {
-  AlertItem,
-  AlertItemWrapper,
-  GetUserAlertsResponse,
+    AlertItem,
+    AlertItemWrapper,
+    GetUserAlertsResponse,
 } from "@/types/api/user";
 
 export default function GlobalHeader() {
-  const [notificationOpen, setNotificationOpen] = useState(false);
-  const { logout, isLogin } = useAuth();
-  const [userInfo, setUserInfo] = useState<{
-    id: string | null;
-    type: string | null;
-    token: string | null;
-  }>({ id: null, type: null, token: null });
+    const [notificationOpen, setNotificationOpen] = useState(false);
+    const { logout, isLogin } = useAuth();
+    const [userInfo, setUserInfo] = useState<{
+        id: string | null;
+        type: string | null;
+        token: string | null;
+    }>({ id: null, type: null, token: null });
 
-  const [notificationQuery, setNotificationQuery] = useState<GetListQuery>({
-    offset: 0,
-    limit: 10,
-  });
+    const [notificationQuery, setNotificationQuery] = useState<GetListQuery>({
+        offset: 0,
+        limit: 10,
+    });
 
-  useEffect(() => {
-    if (isLogin) {
-      const id = localStorage.getItem("id");
-      const token = localStorage.getItem("token");
-      const type = localStorage.getItem("type");
-      setUserInfo({ id, token, type });
-    } else {
-      setUserInfo({ id: null, type: null, token: null });
-    }
-  }, [isLogin]);
+    useEffect(() => {
+        if (isLogin) {
+            const id = localStorage.getItem("id");
+            const token = localStorage.getItem("token");
+            const type = localStorage.getItem("type");
+            setUserInfo({ id, token, type });
+        } else {
+            setUserInfo({ id: null, type: null, token: null });
+        }
+    }, [isLogin]);
 
-  const { data: userNotificationData, refetch: refetchNotification } =
-    useUserAlerts(userInfo.id || "", notificationQuery);
+    const { data: userNotificationData, refetch: refetchNotification } =
+        useUserAlerts(userInfo.id || "", notificationQuery);
 
-  const notifications: AlertItemWrapper[] = userNotificationData?.items || [];
+    const notifications: AlertItemWrapper[] = userNotificationData?.items || [];
 
-  const handleOpenNotifications = () => {
-    setNotificationOpen(true);
-    refetchNotification();
-  };
+    const handleOpenNotifications = () => {
+        setNotificationOpen(true);
+        refetchNotification();
+    };
 
-  const handleLogout = () => {
-    setNotificationOpen(false);
-    logout();
-  };
+    const handleLogout = () => {
+        setNotificationOpen(false);
+        logout();
+    };
 
-  const LinkClasses =
-    "font-bold text-[14px]/[14px] sm:text-[16px]/[20px] whitespace-nowrap";
+    const LinkClasses =
+        "font-bold text-[14px]/[14px] sm:text-[16px]/[20px] whitespace-nowrap";
 
-  return (
-    <header className="w-full h-[70px] px-5 py-[10px]">
-      <div className="flex justify-between sm:justify-center max-w-[1024px] w-full mx-auto sm:gap-8 md:gap-10 items-center">
-        {/* 로고 */}
-        <div>
-          <Link
-            href={"/dashboard"}
-            className="block relative w-[84px] h-[30px] sm:w-[112px] sm:h-[40px] flex-shrink-0"
-          >
-            <Image
-              src={"/assets/images/logo.svg"}
-              fill
-              alt="로고"
-              className="object-contain"
-              priority
-            />
-          </Link>
-        </div>
+    return (
+        <header className="w-full h-[70px] px-5 py-[10px]">
+            <div className="flex justify-between sm:justify-center max-w-[1024px] w-full mx-auto sm:gap-8 md:gap-10 items-center">
+                {/* 로고 */}
+                <div>
+                    <Link
+                        href={"/"}
+                        className="block relative w-[84px] h-[30px] sm:w-[112px] sm:h-[40px] flex-shrink-0"
+                    >
+                        <Image
+                            src={"/assets/images/logo.svg"}
+                            fill
+                            alt="로고"
+                            className="object-contain"
+                            priority
+                        />
+                    </Link>
+                </div>
 
-        {/* 태블릿, PC 검색창 */}
-        <SearchBar className="relative xs:hidden sm:block max-w-[450px]" />
+                {/* 태블릿, PC 검색창 */}
+                <SearchBar className="relative xs:hidden sm:block max-w-[450px]" />
 
-        {/* 네비게이션 */}
-        <nav
-          className={clsx(
-            "flex justify-end items-center gap-[16px] sm:gap-[40px]",
-            { "sm:gap-3 md:gap-10": isLogin }
-          )}
-        >
-          {isLogin ? (
-            <>
-              <Link
-                className={LinkClasses}
-                href={
-                  userInfo.type === "employer"
-                    ? `/shops/${userInfo.id}`
-                    : `/profile/worker/${userInfo.id}`
-                }
-              >
-                {userInfo.type === "employer" ? "내 가게" : "내 프로필"}
-              </Link>
-              <button
-                onClick={handleLogout}
-                className={clsx("block", LinkClasses)}
-              >
-                로그아웃
-              </button>
-              <button
-                onClick={handleOpenNotifications}
-                className="w-5 h-5 relative"
-              >
-                <Image
-                  src={"/assets/icons/notification.svg"}
-                  alt="알림 열기"
-                  fill
-                />
-              </button>
-              {notificationOpen && (
-                <NotificationPanel
-                  notifications={notifications}
-                  onClose={() => setNotificationOpen(false)}
-                />
-              )}
-            </>
-          ) : (
-            <>
-              <Link href={"/login"} className={LinkClasses}>
-                로그인
-              </Link>
-              <Link href={"/signup"} className={LinkClasses}>
-                회원가입
-              </Link>
-            </>
-          )}
-        </nav>
-      </div>
-      <SearchBar className="sm:hidden" />
-    </header>
-  );
+                {/* 네비게이션 */}
+                <nav
+                    className={clsx(
+                        "flex justify-end items-center gap-[16px] sm:gap-[40px]",
+                        { "sm:gap-3 md:gap-10": isLogin }
+                    )}
+                >
+                    {isLogin ? (
+                        <>
+                            <Link
+                                className={LinkClasses}
+                                href={
+                                    userInfo.type === "employer"
+                                        ? `/shops/${userInfo.id}`
+                                        : `/profile/worker/${userInfo.id}`
+                                }
+                            >
+                                {userInfo.type === "employer"
+                                    ? "내 가게"
+                                    : "내 프로필"}
+                            </Link>
+                            <button
+                                onClick={handleLogout}
+                                className={clsx("block", LinkClasses)}
+                            >
+                                로그아웃
+                            </button>
+                            <button
+                                onClick={handleOpenNotifications}
+                                className="w-5 h-5 relative"
+                            >
+                                <Image
+                                    src={"/assets/icons/notification.svg"}
+                                    alt="알림 열기"
+                                    fill
+                                />
+                            </button>
+                            {notificationOpen && (
+                                <NotificationPanel
+                                    notifications={notifications}
+                                    onClose={() => setNotificationOpen(false)}
+                                />
+                            )}
+                        </>
+                    ) : (
+                        <>
+                            <Link href={"/login"} className={LinkClasses}>
+                                로그인
+                            </Link>
+                            <Link href={"/signup"} className={LinkClasses}>
+                                회원가입
+                            </Link>
+                        </>
+                    )}
+                </nav>
+            </div>
+            <SearchBar className="sm:hidden" />
+        </header>
+    );
 }
